fix patrolling

+++ MCB-project/src/subsystems/jetson/AutoAimAndFireCommand.hpp
@@ -55,8 +55,8 @@ private:
 
     int numCyclesForBurst = 0;
     static constexpr int CYCLES_UNTIL_BURST = 200; //cycles
-    static constexpr int BURST_AMOUNT = 0.4; //rad/s
-    static constexpr int PATROL_SPEED = 0.003; //rad/s
+    static constexpr float BURST_AMOUNT = 0.4; //rad/s
+    static constexpr float PATROL_SPEED = 0.003; //rad/s
 
     float yaw = 0.0f, pitch = 0.0f;
     float yawvel = 0.0f, pitchvel = 0.0f;
